+++ src/ChessDotCpp/ChessDotCpp.cpp
@@ -41,7 +41,8 @@ void Test()
     //}
     //magics.close();
 
-    Fen fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
+    //Fen fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
+    Fen fen = "rnbqkbnr/pppppppp/8/8/1P6/8/P1PPPPPP/RNBQKBNR b KQkq - 0 1";
     //Fen fen = "r4rk1/p2n1ppp/3qp3/6B1/N5P1/3P1b2/PPP1BbP1/R2Q1R1K b - - 0 14";
 
     //PerftRunner::Run(fen, 7);
@@ -49,6 +50,24 @@ void Test()
 
     Board board = Board();
     Fens::Parse(board, fen);
+
+    EachColor<Bitboard> pins;
+    PinDetector::GetPinnedToKings(board, pins);
+    Score eval1 = Evaluation::Evaluate(board, pins);
+    
+    auto key = board.Key;
+    board.FlipColors();
+
+    PinDetector::GetPinnedToKings(board, pins);
+    Score eval2 = Evaluation::Evaluate(board, pins);   
+    auto key2 = board.Key;
+    auto fen2 = Fens::Serialize(board);
+
+    board.FlipColors();
+    auto key3 = board.Key;
+    auto fen3 = Fens::Serialize(board);
+    PinDetector::GetPinnedToKings(board, pins);
+    Score eval3 = Evaluation::Evaluate(board, pins);
 
     SearchParameters search_parameters = SearchParameters();
     search_parameters.MaxDepth = Constants::MaxDepth;