+++ src/ChessDotCpp/fen.h
@@ -10,4 +10,6 @@ class Fens
 public:
 	static void Parse(Board& board, Fen fen);
 	static Fen Serialize(const Board& board);
+
+	static void Flip(Board& board, Board& target);
 };